Kerberos: use fewer attributes for g_file_query_info()
We try to preserve the file attributes while glib2 modifies
/etc/krb5.conf by creating an intermediate file. Currently
g_file_query_info() is called so that it will return all file attributes
known be glib2. Not all those attributes can be written back to the new
file which might cause misleading error messages since all attributes
which could be written are written. To avoid this the list of attributes
is restricted to the plain UNIX file permission, SELinux and extendend
attributes.

Resolves: https://gitlab.freedesktop.org/realmd/realmd/-/issues/36