disco: check IPA specific extension in rootDSE
In case an IPA server has restricted anonymous access it might not be
possible to read the based DN to find the domain name or the Kerberos
realm. As long as the rootDSE can be read we can identify an IPA domain
by IPA specific LDAP extensions and derived the domain name from the
default domain suffix. With this patch this information is used as a
fallback if the base DN cannot be read.

Resolves: https://gitlab.freedesktop.org/realmd/realmd/-/issues/44